[FLINK-19414][file connector] Revert "Extract FileSourceTextLinesITCase to AbstractFileSourceITCase" (#13610)
We should not have a test-jar for the flink-connector-files project, because the module it is not scala-versioned. Thus we cannot release an artifact with a dependency on a specific scala version (like the test jar).

We need to have the connector test utils in a different project, or add a scala version to the module.
